Prevent the leak of exceptions from finally block
The task attempts to perform a clean-up by stopping the SonarQube scanner engine from within a finally block. Where the engine is not actually running, as happens when the startup throws an exception, the SonarQube engine throws an exception stating that the engine is not currently running, which subsequently hides the original exception providing the detail into why the engine initially failed to start. This change wraps the `stop` call in a try-catch block that logs the exception and allows execution to continue so any original exceptions are allowed to propagate.