Move to a newer version of the sonar-scanner-api to support API changes
Old versions of sonar-scanner-api attempted to download the URL `batch_bootstrap/index` from the target SonarQube server to retrieve a list of Jars the scanner engine needed to download. This endpoint was deprecated since SonarQube 4.4 and is no longer available in Sonar Cloud meaning versions of the API library that depend on that endpoint no longer work. Migrating to the oldest version of the API that targets the new `bootstrap/index` endpoint overcomes this issues.